% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_test_mf.R
\name{bt_test_mf}
\alias{bt_test_mf}
\title{bt_test_mf}
\usage{
bt_test_mf(
  cs1,
  cs2,
  data,
  subj,
  group = NULL,
  na.rm = FALSE,
  paired = TRUE,
  rscale = "medium",
  phase = "acquisition",
  dv = "scr",
  exclusion = "full data"
)
}
\arguments{
\item{cs1}{cs 1}

\item{cs2}{cs 2}

\item{data}{a data frame containing the dv and iv}

\item{subj}{column name with the participant number.
It should be a unique number.}

\item{group}{name of the group variable, if it is present, default to \code{NULL}}

\item{na.rm}{Whether NAs should be removed, default to \code{FALSE}}

\item{paired}{whether the t-test refers to a paired or independent t-test.}

\item{rscale}{r scale to be used in the prior of the alternative hypothesis}

\item{phase}{Different tests will be run for different phases. That is why
the phase needs to be specified here. Possible values are \code{acquisition},
or \code{acq}, \code{extnction}, or \code{extinction}. See Details for more
information.}

\item{dv}{name of the dependent variable, default to "SCR"}

\item{exclusion}{If any exclusion was done, default to \code{full data}}
}
\value{
a basic function for running a t-test within the \code{multifear}
package
}
\description{
Basic function for running a Bayesian t-test
}
\details{
At the moment the function returns only paired samples t-test. The effect size is Hedge's g.
}
