% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop_cs.R
\name{chop_cs}
\alias{chop_cs}
\title{chop_cs}
\usage{
chop_cs(cs, data, subj, na.rm = FALSE)
}
\arguments{
\item{cs}{index for the CS responses}

\item{data}{Data frame containing the CS columns}

\item{subj}{column nmae with the participant number.
It should be a unique number.}

\item{na.rm}{Should NAs be removed? Default to FALSE.}
}
\description{
Function for separating the conditioned responses into multiple pieces
}
