% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop_css.R
\name{chop_css}
\alias{chop_css}
\title{chop_css}
\usage{
chop_css(cs1, cs2, data, subj, na.rm = FALSE)
}
\arguments{
\item{cs1}{The column name(s) of the conditioned responses for the first conditioned stimulus.}

\item{cs2}{The column name(s) of the conditioned responses for the second conditioned stimulus.}

\item{data}{A data frame containing with all the relevant columns}

\item{subj}{The name of the column including the participant numbers. Unique numbers are expected.}

\item{na.rm}{Whether NAs should be removed, default to \code{FALSE}.}
}
\description{
Function for separating the conditioned responses into multiple pieces for two CSs.
}
\details{
\lifecycle{experimental}
}
