% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_cs.R
\name{combine_cs}
\alias{combine_cs}
\title{combine_cs}
\usage{
combine_cs(cs1, cs2, data, na.rm = FALSE)
}
\arguments{
\item{cs1}{index for the CS1 responses}

\item{cs2}{index for the CS2 responses}

\item{data}{Data frame containing the CS1 and CS2 columns}

\item{na.rm}{Should NAs be removed? Default to FALSE.}
}
\description{
Function for computing mean responses across CSs
}
