% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_cs.R
\name{curve_cs}
\alias{curve_cs}
\title{curve_cs}
\usage{
curve_cs(data)
}
\arguments{
\item{data}{a data frame with the results of a multiverse analyses.}
}
\value{
A plot
}
\description{
Plot the specification curve

\lifecycle{experimental}
}
\details{
This plot is a slight modification of the \code{specr::plot_decisiontree}. Please check the references.
}
\references{
Masur, Philipp K. & Scharkow, M. (2019). specr: Statistical functions for conducting specification curve analyses. Available from https://github.com/masurp/specr.
}
