% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree_cs.R
\name{decision_tree_cs}
\alias{decision_tree_cs}
\title{decision_tree_cs}
\usage{
decision_tree_cs(dat)
}
\arguments{
\item{dat}{a dat frame with the results of a multiverse analyses.}
}
\value{
A plot
}
\description{
Plot decision tree

\lifecycle{experimental}
}
\details{
This plot is a slight modification of the \code{specr::plot_decisiontree}. Please check the references.
}
\references{
Masur, Philipp K. & Scharkow, M. (2019). specr: Statistical functions for conducting specification curve analyses. Available from https://github.com/masurp/specr.
}
