% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Simulated data sets of skin conductance responses}
\format{
A data frame with 40 rows and 22 variables:
\describe{
  \item{CSP1}{Conditioned stimulus plus, trial 1}
  \item{CSP2}{Conditioned stimulus plus, trial 2}
  \item{CSP3}{Conditioned stimulus plus, trial 3}
  \item{CSP4}{Conditioned stimulus plus, trial 4}
  \item{CSP5}{Conditioned stimulus plus, trial 5}
  \item{CSP6}{Conditioned stimulus plus, trial 6}
  \item{CSP7}{Conditioned stimulus plus, trial 7}
  \item{CSP8}{Conditioned stimulus plus, trial 8}
  \item{CSP9}{Conditioned stimulus plus, trial 9}
  \item{CSP10}{Conditioned stimulus plus, trial 10}
  \item{CSM1}{Conditioned stimulus minus, trial 1}
  \item{CSM2}{Conditioned stimulus minus, trial 2}
  \item{CSM3}{Conditioned stimulus minus, trial 3}
  \item{CSM4}{Conditioned stimulus minus, trial 4}
  \item{CSM5}{Conditioned stimulus minus, trial 5}
  \item{CSM6}{Conditioned stimulus minus, trial 6}
  \item{CSM7}{Conditioned stimulus minus, trial 7}
  \item{CSM8}{Conditioned stimulus minus, trial 8}
  \item{CSM9}{Conditioned stimulus minus, trial 9}
  \item{CSM10}{Conditioned stimulus minus, trial 10}
  \item{id}{Participant number}
  \item{group}{Group assignment}
}
}
\usage{
example_data
}
\description{
A simulated data set containing conditioned skin conductance responses for
10 conditioned stimulus plus trials and a equal number of conditioned
stimulus minus trials The variables are as follows:
}
\keyword{datasets}
