% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusion_criteria.R
\name{exclusion_criteria}
\alias{exclusion_criteria}
\title{exclusion_criteria}
\usage{
exclusion_criteria(data, cutoff = 0.01)
}
\arguments{
\item{data}{a data object generated by the \code{multics::chop_css} function}

\item{cutoff}{cut off score. Default to 0.01}
}
\description{
Exclusion criteria
}
\details{
\lifecycle{experimental}

Here the different exclusion criteria are applied to the provided data.
}
