% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_mf.R
\name{forestplot_mf}
\alias{forestplot_mf}
\title{forestplot_mf}
\usage{
forestplot_mf(data, ...)
}
\arguments{
\item{data}{a universe_mf or multiverse_mf object}

\item{...}{any additional argument}
}
\value{
A plot
#'
}
\description{
Basic function for forest plot
}
\details{
This is a wrapper around the \code{forestplot::forestplot} function.
}
