% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_mf.R
\name{mixed_mf}
\alias{mixed_mf}
\title{mixed_mf}
\usage{
mixed_mf(
  cs1,
  cs2,
  data,
  subj,
  time = TRUE,
  group = NULL,
  phase = "acquisition",
  dv = "scr",
  exclusion = "full data"
)
}
\arguments{
\item{cs1}{The column name(s) of the conditioned responses for the first conditioned stimulus.}

\item{cs2}{The column name(s) of the conditioned responses for the second conditioned stimulus.}

\item{data}{A data frame containing with all the relevant columns}

\item{subj}{The name of the column including the participant numbers. Unique numbers are expected.}

\item{time}{should time be included? Default to \code{TRUE}}

\item{group}{the name of the group, if included, default to \code{NULL}.}

\item{phase}{The conditioned phase that the analyses refer to. Accepted values are  \code{acquisition},  or \code{acq}, \code{extinction}, or \code{ext}.}

\item{dv}{name of the measured conditioned response. Default to \code{"SCR"}.}

\item{exclusion}{If any exclusion criteria were applied, default to \code{full data}}
}
\value{
A basic function for running repeated measures ANOVAs
}
\description{
Mixed model function
}
\details{
In case the \code{time} argument is set to true, the function will
include this as a within subjects factor, assuming that the columns in
\code{cs1} and \code{cs2} correspond to ascending time points (e.g., cs1
trial 1, cs1 trial 2 ... cs1 trial \code{n}). If this is not the case, the
results are not to be trusted.

The function uses the \code{ez::ezANOVA} function. The function gives by default a warning regarding the collapsing of factors. This function here suppresses this warning but the user should be aware of it.

The effect size is omega squared.
}
