% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop_cs.R
\name{chop_cs}
\alias{chop_cs}
\title{chop_cs}
\usage{
chop_cs(cs, data, subj, na.rm = FALSE)
}
\arguments{
\item{cs}{index for the CS responses.}

\item{data}{A data frame containing with all the relevant columns}

\item{subj}{The name of the column including the participant numbers. Unique numbers are expected.}

\item{na.rm}{Whether NAs should be removed, default to \code{FALSE}.}
}
\description{
Function for separating the conditioned responses into multiple pieces.
}
\details{
\lifecycle{experimental}
}
