% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop_css.R
\name{chop_css}
\alias{chop_css}
\title{chop_css}
\usage{
chop_css(cs1, cs2, data, subj, na.rm = FALSE)
}
\arguments{
\item{cs1}{index for the CS1 responses}

\item{cs2}{index for the CS2 responses}

\item{data}{Data frame containing the CS columns}

\item{subj}{column nmae with the participant number.
It should be a unique number.}

\item{na.rm}{Should NAs be removed? Default to FALSE.}
}
\description{
Function for separating the conditioned responses into multiple pieces for two CSs
}
