% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_cs.R
\name{combine_cs}
\alias{combine_cs}
\title{combine_cs}
\usage{
combine_cs(cs1, cs2, data, na.rm = FALSE)
}
\arguments{
\item{cs1}{The column name(s) of the conditioned responses for the first conditioned stimulus}

\item{cs2}{The column name(s) of the conditioned responses for the second conditioned stimulus}

\item{data}{A data frame containing all the relevant columns for the analyses}

\item{na.rm}{Whether NAs should be removed, default to \code{FALSE}}
}
\value{
A tibble with the initial data frame (given by the \code{data} argument)
together with an additional column with the means for the columns defined
in the cs1 and cs2 arguments.
}
\description{
Function for computing mean responses across CSs
}
