% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_cs.R
\name{curve_cs}
\alias{curve_cs}
\title{curve_cs}
\usage{
curve_cs(
  data,
  choices = c("exclusion", "cut_off"),
  labels = c("A", "B"),
  rel_heights = c(2, 3),
  null = 0,
  ci = TRUE,
  ribbon = FALSE,
  sample_perc = 1
)
}
\arguments{
\item{data}{a data frame with the results of a multiverse analyses}

\item{choices}{a vector specifying which analytical choices should be plotted}

\item{labels}{labels for the two parts of the plot}

\item{rel_heights}{vector indicating the relative heights of the plot}

\item{null}{Indicate what value represents the 'null' hypothesis (defaults to
zero).}

\item{ci}{logical value indicating whether confidence intervals should be
plotted.}

\item{ribbon}{logical value indicating whether a ribbon instead should be
plotted.}

\item{sample_perc}{numeric value denoting what percentage of the
specifications should be plotted. Needs to be strictly greater than 0 and smaller than 1.
Defaults to 1 (= all specifications). Drawing a sample from all
specification usually makes only sense of the number of specifications is
very large and one wants to simplify the visualization.}
}
\value{
A plot
}
\description{
Plot the specification curve

\lifecycle{experimental}
}
\details{
This plot is a slight modification of the \code{specr::plot_specs} function. Please check the references and credit the original authors.
}
\references{
Masur, Philipp K. & Scharkow, M. (2019). specr: Statistical functions for conducting specification curve analyses. Available from https://github.com/masurp/specr.
}
