% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusion_criteria.R
\name{exclusion_criteria}
\alias{exclusion_criteria}
\title{exclusion_criteria}
\usage{
exclusion_criteria(data, cutoff = 0)
}
\arguments{
\item{data}{a data object generated by the \code{multics::chop_css} function}

\item{cutoff}{cut off score}
}
\value{
A tibble with a brace of smaller tibbles, with each sub-tibble including a data frame after each one of the exclusion criteria -- mentioned in the \code{multifear::chop_css} function -- is applied.
}
\description{
Exclusion criteria
}
\details{
Here the different exclusion criteria are applied to the provided data.
}
