% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_mf.R
\name{forestplot_mf}
\alias{forestplot_mf}
\title{forestplot_mf}
\usage{
forestplot_mf(
  data,
  ci = TRUE,
  include_label_text = TRUE,
  reported_effect = "common",
  ...
)
}
\arguments{
\item{data}{a universe_mf or multiverse_mf object}

\item{ci}{should confidence intervals be included -- default to TRUE}

\item{include_label_text}{Whether the labels for each effect should be include. Default to \code{TRUE}}

\item{reported_effect}{Option as to what effect size to report. The options are: "common" (default option), "anova", or "ttest".}

\item{...}{any additional argument}
}
\value{
A plot
#'
}
\description{
Basic function for forest plot
}
\details{
This is a wrapper around the \code{forestplot::forestplot} function.
The function only uses the ANOVAs and the t-tests. For the t-tests though
it includes only the two-sided, as they are the same with the one-sided ones
and having both would probably give a false picture of the effect.

The include_label_text can be used when multiple plots need to be combined
side by side, so one of them does not include the labels.
}
