% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_cs.R
\name{inference_cs}
\alias{inference_cs}
\title{inference_cs}
\usage{
inference_cs(data, alpha_level = 0.05, na.rm = FALSE)
}
\arguments{
\item{data}{a tibble with the results of a multiverse analyses}

\item{alpha_level}{What should be the alpha level used (default to 0.05)}

\item{na.rm}{Should NAs be removed? Default to FALSE}
}
\value{
A data frame with results
}
\description{
Function for summarizing the multiverse results
}
\details{
For now the function returns only mean p values and proportion of p values below a criterion defined by the \code{alpha_level} parameter (default to 0.05).
}
