% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_cs.R
\name{inference_cs}
\alias{inference_cs}
\title{inference_cs}
\usage{
inference_cs(data, alpha_level = 0.05, na.rm = FALSE, framework = "Both")
}
\arguments{
\item{data}{a data frame with the results of a multiverse analyses.}

\item{alpha_level}{What should be the alpha level used (default to 0.05).}

\item{na.rm}{Should NA's be removed (default to \code{FALSE}). See details.}

\item{framework}{Inference framework. Values could be "NHST", "Bayesian", or "Both" (no case sensitivity)}
}
\value{
A data frame with summaries of the results.
}
\description{
Function for summarizing the multiverse results.
}
\details{
For now the function returns mean, median, standard deviations of p values and proportion of p values below a criterion defined by the \code{alpha_level} parameter (default to 0.05) as well as mean Bayes factors (please see the `framework` argument). The user may choose to drop the NAs for the summary statistic.
}
