% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_mf.R
\name{mixed_mf}
\alias{mixed_mf}
\title{mixed_mf}
\usage{
mixed_mf(
  cs1,
  cs2,
  data,
  subj,
  group = NULL,
  phase = "acquisition",
  dv = "scr",
  exclusion = "full data"
)
}
\arguments{
\item{cs1}{The column name(s) of the conditioned responses for the first conditioned stimulus.}

\item{cs2}{The column name(s) of the conditioned responses for the second conditioned stimulus.}

\item{data}{A data frame containing with all the relevant columns}

\item{subj}{The name of the column including the participant numbers. Unique numbers are expected.}

\item{group}{the name of the group, if included, default to \code{NULL}.}

\item{phase}{The conditioned phase that the analyses refer to. Accepted values are  \code{acquisition},  or \code{acq}, \code{extinction}, or \code{ext}.}

\item{dv}{name of the measured conditioned response. Default to \code{"SCR"}.}

\item{exclusion}{If any exclusion criteria were applied, default to \code{full data}}
}
\value{
A data frame with the results.

The data frame returned is the standard one returned in all function in the package (see e.g., \code{multifear::rm_anova_mf}). The important difference is of course the `x` column. There, you can see the term of the model that is returned. So, not the full model is returned but only this particular term.
}
\description{
Basic function for running mixed models for the multiverse analysis
}
\details{
\lifecycle{experimental}

The function assumes that you include more than 1 trial per CS. The function returns an error if that is not the function.

The function performs by default two dependent variable standardizations, the one per subject
and the other one without taking subject into account.

In case time is included, the function computes the intercept -- i.e., the 0 point -- on the middle of the time sequence.

The following models are run and compared: a) Intercept only model, b) Intercept plus CS model, and c) Intercept plus CS \code{x} Time interaction.

Separate models are run with `Subject` as random factor, as well as `Subject and Time` as random factors.

The model is fit by maximizing the log-likelihood (i.e., "ML" term in nlme::lme).

The model comparison is done using `BIC`.
}
\seealso{
\code{\link[nlme]{lme}}
}
