% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_cs.R
\name{multi_cs}
\alias{multi_cs}
\title{multi_cs}
\usage{
multi_cs(
  cs1,
  cs2,
  data,
  subj,
  group = NULL,
  phase = "acquisition",
  na.rm = FALSE,
  print_output = TRUE
)
}
\arguments{
\item{cs1}{cs 1}

\item{cs2}{cs 2}

\item{data}{a data frame containing the dv and iv}

\item{subj}{column nmae with the participant number.
It should be a unique number.}

\item{group}{name of the group variable, if it is present (default to NULL)}

\item{phase}{Different tests will be run for different phases. That is why
the phase needs to be specified here. Possible values are \code{acquisition},
or \code{acq}, \code{extinction}, or \code{extinction}. See Details for more
information.}

\item{na.rm}{Should NAs be removed? Default set the \code{FALSE}}

\item{print_output}{Whether to print the output or not. Default set to \code{TRUE}}
}
\description{
Basic function for conducting multi-verse analyses of conditioning
data
}
\details{
In case of higher order interaction, only the highest order
effect is shown
}
