% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse_cs.R
\name{multiverse_cs}
\alias{multiverse_cs}
\title{multiverse_cs}
\usage{
multiverse_cs(
  cs1,
  cs2,
  data,
  subj,
  group = NULL,
  cs_paired = NULL,
  include_bayes = TRUE,
  phase = "acquisition",
  cutoff = c(0, 1, 2, 3),
  print_output = TRUE
)
}
\arguments{
\item{cs1}{The column name(s) of the conditioned responses for the first conditioned stimulus}

\item{cs2}{The column name(s) of the conditioned responses for the second conditioned stimulus}

\item{data}{A data frame containing with all the relevant columns}

\item{subj}{The name of the column including the participant numbers. Unique numbers are expected.}

\item{group}{the name of the group, if included, default to \code{NULL}.}

\item{cs_paired}{A character vector with the trials that were paired. Default is set to \code{NULL}, suggesting that there was full reinforcement}

\item{include_bayes}{Whether the bayesian analyses should be run. Default to \code{TRUE}}

\item{phase}{The conditioned phase that the analyses refer to. Accepted values are  \code{acquisition},  or \code{acq}, \code{extinction}, or \code{ext}.}

\item{cutoff}{A numeric vector of the cutoff criteria applied. Default to \code{0, 0.05, .1}}

\item{print_output}{Whether to print the output or not. Default set to \code{TRUE}}
}
\value{
A tibble with the following column names:
x: the name of the independent variable (e.g., cs)
y: the name of the dependent variable as this defined in the \code{dv} argument
exclusion: see \code{exclusion} argument
model: the model that was run (e.g., t-test)
controls: ignore this column for this test
method: the model that was run
p.value: irrelevant here
effect.size: irrelevant here
estimate: the estimate of the test run
statistic: the t-value
conf.low: the lower confidence interval for the estimate
conf.high: the higher confidence interval for the estimate
data_used: a list with the data used for the specific test.

#' # Load data and define argument to be fed in universe_cs
data("example_data", package = "multifear")
example_data <- example_data[1:10, ]
cs1 <- paste0("CSP", 1:10)
cs2 <- paste0("CSM", 1:10)
subj = "id"
multiverse_cs(cs1, cs2, subj = subj, data = example_data)
}
\description{
Basic function for conducting multiverse analyses of conditioning
data
}
\details{
\lifecycle{experimental}

In case of higher order interaction, only the highest order
effect is returned.
}
