% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse_cs.R
\name{multiverse_cs}
\alias{multiverse_cs}
\title{multiverse_cs}
\usage{
multiverse_cs(
  cs1,
  cs2,
  data,
  subj,
  group = NULL,
  include_bayes = TRUE,
  cutoff = c(0, 0.1, 0.05),
  phase = "acquisition",
  print_output = TRUE
)
}
\arguments{
\item{cs1}{The column name(s) of the conditioned responses for the first conditioned stimulus.}

\item{cs2}{The column name(s) of the conditioned responses for the second conditioned stimulus.}

\item{data}{A data frame containing with all the relevant columns}

\item{subj}{The name of the column including the participant numbers. Unique numbers are expected.}

\item{group}{the name of the group, if included, default to \code{NULL}.}

\item{include_bayes}{Whether the bayesian analyses should be run. Default to \code{TRUE}.}

\item{cutoff}{A numeric vector of the cutoff criteria applied. Default to \code{0, 0.05, .1}}

\item{phase}{The conditioned phase that the analyses refer to. Accepted values are  \code{acquisition},  or \code{acq}, \code{extinction}, or \code{ext}.}

\item{print_output}{Whether to print the output or not. Default set to \code{TRUE}.}
}
\description{
Basic function for conducting multiverse analyses of conditioning
data
}
\details{
\lifecycle{experimental}

In case of higher order interaction, only the highest order
effect is returned.
}
