% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_anova_mf.R
\name{rm_anova_mf}
\alias{rm_anova_mf}
\title{rm_anova_mf}
\usage{
rm_anova_mf(
  cs1,
  cs2,
  data,
  subj,
  time = TRUE,
  group = NULL,
  phase = "acquisition",
  dv = "scr",
  exclusion = "full data"
)
}
\arguments{
\item{cs1}{cs 1}

\item{cs2}{cs 2}

\item{data}{a data frame containing the dv and iv}

\item{subj}{column nmae with the participant number.
It should be a unique number.}

\item{time}{should time be included? Default to \code{TRUE}}

\item{group}{name of the group variable, if it is present, default to \code{NULL}}

\item{phase}{Different tests will be run for different phases. That is why
the phase needs to be specified here. Possible values are \code{acquisition},
or \code{acq}, \code{extinction}, or \code{extinction}. See Details for more
information.}

\item{dv}{name of the dependent variable, default to "SCR"}

\item{exclusion}{If any exclusion was done, default to \code{full data}}
}
\value{
A basic function for running repeated measures ANOVAs
}
\description{
Basic function for running repeated measures ANOVA
}
\details{
In case the \code{time} argument is set to true, the function will
include this as a within subjects factor, assuming that the columns in
\code{cs1} and \code{cs2} corrrespond to ascending time points (e.g., cs1
trial 1, cs1 trial 2 ... cs1 trial \code{n}). If this is not the case, the
results are not to be trusted.

The function uses the \code{ez::ezANOVA} function. The function gives by default a warning regarding the collapsing of factors. This function here suppresses this warning but the user should be aware of it.

The effect size is omega squared.
}
