% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_mf.R
\name{t_test_mf}
\alias{t_test_mf}
\title{t_test_mf}
\usage{
t_test_mf(cs1, cs2, data, paired = TRUE, phase = "acquisition")
}
\arguments{
\item{cs1}{cs 1}

\item{cs2}{cs 2}

\item{data}{a data frame containing the dv and iv}

\item{paired}{whether the t-test refers to a paired or independent t-test.}

\item{phase}{Different tests will be run for different phases. That is why
the phase needs to be specified here. Possible values are \code{acquisition},
or \code{acq}, \code{extnction}, or \code{extinction}. See Details for more
information.}
}
\value{
a basic function for running a t-test within the \code{multifear}
package
}
\description{
Basic t.test function for running a t-test
}
