% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe_cs.R
\name{universe_cs}
\alias{universe_cs}
\title{universe_cs}
\usage{
universe_cs(
  cs1,
  cs2,
  data,
  subj,
  group = NULL,
  phase = "acquisition",
  na.rm = FALSE,
  dv = "scr",
  print_output = TRUE,
  exclusion = "full data"
)
}
\arguments{
\item{cs1}{The column name(s) of the conditioned responses for the first conditioned stimulus.}

\item{cs2}{The column name(s) of the conditioned responses for the second conditioned stimulus.}

\item{data}{A data frame containing with all the relevant columns}

\item{subj}{The name of the column including the participant numbers. Unique numbers are expected.}

\item{group}{the name of the group, if included, default to \code{NULL}.}

\item{phase}{The conditioned phase that the analyses refer to. Accepted values are  \code{acquisition},  or \code{acq}, \code{extinction}, or \code{ext}.}

\item{na.rm}{Should NAs be removed? Default set the \code{FALSE}}

\item{dv}{name of the measured conditioned response. Default to \code{"SCR"}.}

\item{print_output}{Whether to print the output or not. Default set to \code{TRUE}}

\item{exclusion}{If any exclusion was done, default to \code{full data}}
}
\description{
Basic function for conducting universe analyses of conditioning
data
}
\details{
In case of higher order interaction, only the highest order
effect is shown
}
